pyproject.toml
[project]
name = "openmare"
version = "0.1.0"
description = "Adalah pokoknya..."

author = "https://github.com/zidny-z/"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "openai>=2.14.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.10",
]
[tool.ruff.format]
indent-style = "tab"
